Update patch set 2
Patch Set 2:

> Patch Set 2:
> I indeed did that change locally, but while reviewing I realized that it would possibly make more sense to let the overall preparation not fail in such cases at all. The wording of the message at least suggests that it was meant to be ignored.

The spec says that it is, yes?

> It is a tradeoff that could also be determined by numbers, i.e. how many interrupts are routed vs. how many were assigned by the firmware. If there are more routed interrupts with the IO-APIC than without, it should probably be used.

Well, maybe it should, but any item that sets it as the interrupt line in PCI config space is, per the spec, invalid. So even if it's a valid line, anything that says it is on that line is incorrect and should be ignored.

> We must also ensure that the 0xff interrupt line is reset to 0 in such cases, as otherwise the exact problem the Linux commit tries to avoid may come up (i.e. a driver trying to use 255 as the interrupt). I have not checked whether this already happens implicitly. The reason why the Linux commit only happened in 2016 is probably that before the widespread use of MSIs, it was extremely unlikely that you would collide on such a high numbered interrupt.

All drivers I checked so far seem to be aware of this already: http://xref.plausible.coop/source/xref/haiku/src/add-ons/kernel/busses/usb/xhci.cpp#839

And this is only for x86; it seems that 255 is a valid line on non-x86, at least for now (though that spec is from 2005, I didn't try to locate a newer one.)

> Please also adjust the debug output to print the actual value, so that we can tell the cases apart without having to look up the real value in the PCI dump.

OK, I'll do that.

> I've seen that the system in #14504 is also impacted by this (there it is an unrouted SMBus device). So that one should be retested as well.

OK.

Patch-set: 2