Update patch set 2
Patch Set 2:

> Patch Set 2:
> > Patch Set 2:
> > I indeed did that change locally, but while reviewing I realized that it would possibly make more sense to let the overall preparation not fail in such cases at all. The wording of the message at least suggests that it was meant to be ignored.
> 
> The spec says that it is, yes?

No, I meant the actual B_ERROR return that was previously hit. The wording of the message suggests that it should not return B_ERROR at all (which actually aborts preparation and prevents the use of IO-APICs). But looking at the commit history the wording was changed in d8ce1a9 to "solve" a panic, without actually understanding what the B_ERROR return would lead to. So the proposal would be to fix up d8ce1a9 to not only making the panic into a dprint, but actually "show warning and keep going" as the commit message suggests by removing the B_ERROR return.

> > It is a tradeoff that could also be determined by numbers, i.e. how many interrupts are routed vs. how many were assigned by the firmware. If there are more routed interrupts with the IO-APIC than without, it should probably be used.
> 
> Well, maybe it should, but any item that sets it as the interrupt line in PCI config space is, per the spec, invalid. So even if it's a valid line, anything that says it is on that line is incorrect and should be ignored.

This is the same misunderstanding as above. The interrupt routing preparation is global across all devices/interrupts in the system, not just for this specific device. Returning an error from here will completely disable IO-APICs which affects all devices. In case there is a serious mismatch (systems regularly used to have messed up ACPI tables that would result in these) then enabling IO-APICs with the bogus data would render the system broken, as many devices would end up without interrupt routing. In such cases it makes sense to fail the preparation step. In cases where the firmware configured only the minimal set of interrupts and the interrupt routing information is (mostly) valid, it makes a lot of sense to use IO-APICs even if some devices still end up without routing (because more devices end up routed that without IO-APICs).

> > We must also ensure that the 0xff interrupt line is reset to 0 in such cases, as otherwise the exact problem the Linux commit tries to avoid may come up (i.e. a driver trying to use 255 as the interrupt). I have not checked whether this already happens implicitly. The reason why the Linux commit only happened in 2016 is probably that before the widespread use of MSIs, it was extremely unlikely that you would collide on such a high numbered interrupt.
> 
> All drivers I checked so far seem to be aware of this already: http://xref.plausible.coop/source/xref/haiku/src/add-ons/kernel/busses/usb/xhci.cpp#839

That's not a really good reason not to ensure that it is safe in the central place where things go through. The driver ideally should not have to have these checks, especially as they are platform specific.

> And this is only for x86; it seems that 255 is a valid line on non-x86, at least for now (though that spec is from 2005, I didn't try to locate a newer one.)

The interrupt lines are already rewritten when using IO-APICs, because the interrupt numbers are different. So ensuring that the invalid number is reset to 0 at that point makes the most sense.

> > Please also adjust the debug output to print the actual value, so that we can tell the cases apart without having to look up the real value in the PCI dump.
> 
> OK, I'll do that.

I see you already did that. It could be made cleaner though as you used a uint32 for a value that is only a uint8 and used an B_PRId32 for an unsigned type.

Patch-set: 2