Update patch set 1
Patch Set 1:

CD boot is needed rarely enough we don't need to make it our first priority. I intend to add a "no sources" release profile which does fit on a CD, but this will of course only be downloadable from the website so that we can keep our current GPL compliance model.

Patch-set: 1