Update patch set 3
Patch Set 3: Code-Review+2

The code looks correct now. Just one observation: Could the logging function be expected to return the BPositionIO into its previous state? There are now two places in the code which need to seek back after logging, but it seems like the situation with logging would always be like this, you'd always want it to not have a side-effect. So the remembering of the current position and seeking back to it could be moved into the logging function. Do you see anything that speaks against this? In any case, the patch is certainly fine as is.

Patch-set: 3
Reviewer: Gerrit User 1000024 <1000024@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2