Update patch set 1
Patch Set 1:

Thank you for the links; the patch is avoiding the recalculation of the size of the BString and also the expensive (owing to complex pluralisation) generation of the localized text that goes into the View.  In the case of PecoRename's "StatusView" this process would still be happening so I don't think this will fix the problem?

Patch-set: 1