Update patch set 1
Patch Set 1:

Previously the "_GetLabel" method was rendering the test using the "BStringFormat".  When the count changed, the "MinSize" was being invoked each time which was +1 call.  If, as a result, the layout was being invalidated then the logic would presumably call MinSize and MaxSize again which would be +2 calls.  Next the view itself was invalidated and required repainting which was another +1 call.  So in total it was needing to perform this 'string generation' process two to four times for each change of the count.  Now it will be doing this once with the downside of having a quasi fixed width.

Patch-set: 1