Update patch set 1
Patch Set 1:

Hi Stefan; the comment is a warning that there is a speed issue in this method.  I think it is OK and is only confusing in the context of this current change.  Yes, I had noted on the ticket that a better solution would be to only render the label every second or so.  I had hoped there was a 'SendMessage(...)' that would take a delay or something, but without that it's probably not worthwhile for now because this solution is fast enough for the time being.

Patch-set: 1
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0