Update patch set 1
Patch Set 1:

Another perspective on this, is that it makes it easier to compile foreign software on Haiku. It will decrease the number of platform specific patches that are necessary to compile. 

The nature of the missing functions is such that it is not a disaster that they do not work as expected. A partially broken thread-priority setting should not be a software breaking bug.

I don't have a firm opinion on whether or not to include this patch. I just wanted to give another perspective.

Patch-set: 1
Reviewer: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0