Update patch set 2
Patch Set 2: Code-Review+1

The overall change looks good, thanks!

I'm not too fond of the 'msgIter' and 'thisFetch' names, I'd prefer 'iterator' and 'uidsChunk' or something like that.

Besides that, it could take the chance and improve the error handling in case AddItem() fails -- that was poor on my part, and got slightly worse with this change (it should break out when an error occurs, log it, and stop proceeding).

Finally, I'd move the command creation and enqueing into its own private method.

Patch-set: 2
Label: Code-Review=+1