Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> > If you can't be bothered to fix your own mistakes, you can at least
>  > fix style issues by example, rather than critque against weakly
>  > constructed standards you don't entirely follow consistently
>  > yourself, or not nitpick over acceptable patches in the first
>  > place.

We're in fact very picky about the coding style, sorry for that. There are are plenty of good reasons to be picky about that, though.

> Cases in point:
> 
> * line 460 - indentation is per the coding guidelines as function arguments require a newline break to not exceed column 80, and the *standard as demonstrated* is to indent one past the indent-aligned method name.  No other instance in this file complies, but no other instance wraps either - given the lack of a suitable example to follow (voiding the "follow existing code" argument), I have followed the coding guidelines as written.

The coding style does not even mention how inline methods are written; that's certainly not your fault, but this still doesn't mean that those rules don't exist -- there are plenty of examples in our codebase.

When in doubt, the best way is to reproduce the style you find in the file you're working on.

> * with respect to ordering of methods vs members - this is not in the coding style guide.

Yes, that's not mentioned either. The example in the guide adheres to this, though. Anyway, this just means that the guide needs a patch, too.

> * line 464 is a deficiency in your design, and not a fault of the bug-fix for your existing design or a complaint of style.

It's an implementation that closely follows the needs. Since I did not take a line limit into account (and that's the deficiency), there was no point in passing iterators. The corrected requirements change this, of course.

A bug fix should be taken as a chance to improve the code quality. If you're not interested in fixing my bugs, I wonder why you wrote that patch in the first place.

Anyway, it's all yours to decide how to proceed. I'm sure that someone may rework that patch to be accepted some day.

Patch-set: 3