Update patch set 2
Patch Set 2:

> IMO, instead of _BTabData_, we should instead make BPrivate::BTab the class with private data. This is more like what Qt does here, and would allow for a more robust solution to some other ABI issues down the line, should we adopt this paradigm elsewhere.

We still need reserved space for virtuals, anyway. As long as we have lots of space left, why not put it to use before introducing a kludge, that you can introduce later at any point, anyway?

I'm not too fond of the Qt solution, and it can and should only be used with stuff that is actually private. It doesn't solve any ABI issues. It just makes the IMO ugly last resort solution the default one.

> Also, please update the TabView documentation with this change.

That would be great, indeed.

Patch-set: 2