Update patch set 2
Patch Set 2:

> Patch Set 1: Code-Review-1
> 
> The Deskbar changes aren't complete, of course; the kMaxReplicantHeight constant will be removed completely. I just want to know if everyone is okay with the extension of the instantiate_deskbar_item() call. Old add-ons will still work (just with a small icon), new ones won't work on an older Deskbar, though.

I'm okay with this change, but how will older application still work, if we are looking for a different function signature?
Is it because it's a C function, so it's not mangled differently, even if parameters are different?

Patch-set: 2
Reviewer: Gerrit User 1000062 <1000062@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0