Update patch set 1
Patch Set 1:

Tested that it still boots on my machine but that's it. I'd like to confirm there was actually a bug in this now, and the 4TB drive I have doesn't work yet.

Also I think it would be better to set the partition block size from the device size at first, and then use the partition block siee.

Effects on ramdisks and mounted image files must be reviewed too (they dont use 512 bytes sectors and are not devices, respectively)

Patch-set: 1
Label: Code-Review=0