ffmpeg: Use ffmpeg-provided timestamp
It makes no sense to use the packet dts as a presentation timestamp.
However, the frames don't reliably have a dts. ffmpeg provides us with a
"best effort" timestamp which is more likely to be correct (or as
correct as possible) in all cases.

Note that this relies on the previous commit to get the relevant
information from the AV Packet.

While I'm at it, also convert the timestamp to microseconds as required
by the media_header struct, and don't assume anything about the ffmpeg
timebase.