Update patch set 1
Patch Set 1:

I don't intend to change this actually; the code is a mess, and the look is still acceptable. If John wants to improve upon it, I wish him well, though :-)

The change in ExpandoMenuBar is in line with this -- it can stay this way until we change the width of the leaf menu in that mode. To fix this correctly, the complete resizing and layout mechanism would need to be replaced. Personally, I'd rather start with Deskbar2 instead of sinking any time there.

The leaf is in the repository as a WonderBrush file (data/artwork/HAIKU Leaf).

Patch-set: 1