Update patch set 3
Patch Set 3:

But it is for the one place... We do not need this class anywhere...
I am in doubt. May be someone can beautify patch then, when there will be more such places?... And i even don't know where to put class definition...

Anyway, let's someone decide what to do. In my opinion, it is better to leave patch as is. Who can decide it? Have i freedom to chose solution? :)

Patch-set: 3