Update patch set 3
Patch Set 3:

The problem that we might have a different point of views on "beauty" and i do not have criteria to make decision myself.

In my point of view, the best solution is: break + status out of the {}. But i am understand that it is C-style. Or current variant (it is not so bad in fact, because it is localized).

So, i don't know how to react. Because i have no critera. 
The variant with class - a lot of effort for simple fix.
The vatian with if (status == B_OK -- adds one more level of ident. In my opinion it is much worse.

There are so many possible solutions...

So, who can make final decision? Just tell it, i will try to code even i dislike it.

Another variant. We closing this patch and i will add bugreport in bugtracker. Someone who can, will write brilliant code. :)

Patch-set: 3
Reviewer: Gerrit User 1000072 <1000072@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0