Update patch set 1
Patch Set 1:

What was merged is the "while I'm at it" part of the commit message. Basically we were just putting the dts (decoding timestamp, expressed in stream units, usually frames) into the media header presentation time (which should be microseconds).

So, half the nonsense is fixed: we still use a decoding timestamp as the presentation timestamp, but at least we use it with the correct unit.

The remaining part of this patch is to use the ffmpeg computed best_effort_timestamp (that will be the pts when available, and an estimation from the dts and frames durations otherwise). However, for this to work we need the full timestamp info from the ffmpeg codec, which is what the previous commit is about.

Patch-set: 1