Update patch set 1
Patch Set 1:

To clarify why this is WIP:
currently we just store a pointer to the AVPacket into the media header extra data. There is no guarantee that the AVPacket will still be allocated (and still hold the right data) by the time this gets to the decoder side.

We need to either get ffmpeg reference counting to work as expected on the packet, or make a copy of it on the decoder side. However, it's more complicated than this, because there is then a transfer of ownership from the reader to the decoder, and we can't be sure they are used together (ffmpeg to read a stream + some other codec to decode the data is possible in our architecture). If this was to happen we would get a memory leak then.

Patch-set: 1
Label: Code-Review=0