mail: add B_MAIL_ATTR_READ == B_UNREAD attr for New emails fetched by IMAP
Currently, there are two attributes: B_MAIL_ATTR_READ and B_MAIL_ATTR_STATUS.
B_MAIL_ATTR_READ is set in POP3 inbound add-on.
B_MAIL_ATTR_STATUS can be set in IMAP inbound add-on.
If there is no IMAP mapped status, the "New" status is set, but no read attribute.

So, the new e-mails are sometimes marked with read attribute and sometimes -- not.

This commit fixes this situation.