Update patch set 4
Patch Set 4:

Code looks fine now, but I have tested this change and making fonts bigger is fairly annoying. The bigger the font gets, the wider the window is, and "+" is running away from the cursor.
Please make the width static, so the window doesn't get wider, and use BFont::TruncateString for bigger sizes.

It would be great if you could add an empty line before Change-Id in commit message, but that's a minor nitpick.

Patch-set: 4