Update patch set 4
Patch Set 4:

> Code looks fine now, but I have tested this change and making fonts
 > bigger is fairly annoying. The bigger the font gets, the wider the
 > window is, and "+" is running away from the cursor.
 > Please make the width static, so the window doesn't get wider, and
 > use BFont::TruncateString for bigger sizes.
 > 
 > It would be great if you could add an empty line before Change-Id
 > in commit message, but that's a minor nitpick.

I agree that it is annoying. I dunno how to solve this the best way. Even if I set the ExplicitMaxSize and B_TRUNCATE_END on BStringView it wants to resize. Maybe I missing something here, or it is a bug in BStringView. If the user sets the font to 24 pt and we needs to truncate the preview, next time the user starts the application we want it to look normal (otherwise the window would look very small with its current font settings). Any ideas?

Patch-set: 4
Label: Code-Review=0