Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> That's why I am proposing in a hypotetical future API to remove the correspondence Looper <-> Thread and allow for a more flexible pattern where loopers can for example run in pools of threads, so the user has the choice of the threading model instead of being tied to this.

Huh? What is the value of that? It would make locking with loopers involved MUCH more complicated for no real apparent benefit. If you want to receive BMessages without a looper, we may be able to facilitate that, but decoupling BLooper from a thread makes no sense.

Patch-set: 1