Update patch set 1
Patch Set 1: Code-Review+1

This is very similar to how Qt handles event loops, and that seems to work well.

It would be nice to be able to stop the looper from deleting itself on quit. Then one could use this API to create windows that were not on a separate thread, for instance, among other things. But I guess that's outside the scope of this change.

Patch-set: 1
Label: Code-Review=+1