Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > > Huh? What is the value of that? It would make locking with loopers involved MUCH more complicated for no real apparent benefit. If you want to receive BMessages without a looper, we may be able to facilitate that, but decoupling BLooper from a thread makes no sense.
> > 
> > No it would make developers more free to choose their threading paradigm.
> 
> Well, the threading paradigms it would make possible make no sense at all, and would be incredibly confusing to deal with. Running a BLooper on top of an existing thread makes sense, which this commit makes possible; receiving messages without a full BLooper may make sense, but otherwise, adding "thread pools" or the like would just make cross-thread synchronization and data sharing unbelievably confusing.

Okay, I understand putting a consideration at random can be misunderstood easily. It depends on the paradigm you use behind. Something like goroutines and channels makes this indeed possible. One confunsing thing for example is having BLooper as a BHandler subclass, messages handling should be probably put only in the latter. But that's another story and this is probably the wrong place to discuss it.

Patch-set: 1