Update patch set 1
Patch Set 1:

As a big mailing list user, I find the "To" field quite important to know if the mail is addressed directly to me, or to some mailing list.

Would it be possible to have From and To on the same line, so as to not waste too much space?

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0