Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> As a big mailing list user, I find the "To" field quite important to know if the mail is addressed directly to me, or to some mailing list.

Good point.

> Would it be possible to have From and To on the same line, so as to not waste too much space?

I'm sure it's possible, though that may have other issues. If the mail addresses of From an/or To are rather long, you ma not be able to have as narrow a window as you like. Truncating the To-field may be possible (I'm not sure where/how that exactly has to be done though. It's also kinda ugly IMO...

I'd rather keep the To-field in a separate line, I think.

Patch-set: 1
Label: Code-Review=0