Update patch set 2
Patch Set 2:

Hi Peter,

I've looked a bit through the code, and I cannot see where an IMAPConnectionWorker instance is actually deleted. The destructor would delete the semaphore. Of course when the worker is right in the middle of executing a command and also (potentially) excesses its members while doing so, the IMAPProtocol cannot simply delete the workers to prevent them from blocking on the semaphore. But anyhow, where are the workers deleted? My intuition is that the worker class should get a Quit() method or something, where the semaphore is deleted. It should not happen in the destructor. And then when the worker thread exits, it selfdestructs. That way, we can be sure that no memory is freed too early. Maybe your solution is fine, but the shutdown sequence takes longer than it could, and also the wake once per second is not nice. What do you think?

Patch-set: 2
Label: Code-Review=0