Update patch set 3
Patch Set 3:

> Patch Set 3:
> However, this allows shortcuts like "é". Do we really want that? They may not be composable at all on some keyboards (eg. spanish has accents on dead keys).

Probably not; as this will also allow "composed" characters from ideogram-based writing systems (Chinese, Japanese, etc.)

Maybe we want 3 passes: a first that accepts uppercase Latin characters, a second that accepts any Latin characters, and finally one that accepts all Unicode characters.

But even this is a bad heuristic, because there are plenty of non-Latin character sets that are phonetic and do not require an input method, e.g. Cyrillic. Ideally, we would be able to check if a character was type-able given a user's keymap -- though this raises the additional question of 'what if they have keymap switcher installed'. But this may be too resource-intensive to do within BMenu? (Perhaps there should be a way to query the input_server for this information?)

Patch-set: 3