Update patch set 3
Patch Set 3:

Well, the app developer can set their own shortcuts, and this is only a fallback when they didn't. So I guess it's better than nothing.

We should probably review the whole shortcut system to be more locale-neutral, but that's a lot more complex. When you use a non-latin keyboard, all common shortcuts (alt+c alt+v etc) are not reachable anyways, so we have way more important things to fix.

Patch-set: 3