Update patch set 3
Patch Set 3:

(2 comments)

I opted to leave spaces instead of tabs in the "-h" fprintf code. Switching to tabs causes display issues. It seems like Terminal, or maybe bash renders tabs as 8sp. With tabs, the indentation is wrong between lines. I've been editing with Pe, set to the stock 4sp tabs. When I check the code (using tabs instead of spaces) in vim w/8sp tabs, the indentation is the same as if I ran Terminal -h from the shell. Keeping spaces in the code, as originally written, keeps both the code and runtime readability in check.  I'll look into it more, but feel it should be a separate issue. I improved the comments to reflect this.

Patch-set: 3