Update patch set 1
Patch Set 1:

Sure, here's the report: https://gist.github.com/waddlesplash/c51b7e3bfccd001c85594274c0588b51

As mentioned in the commit message, it's an assert failure, not a generic segfault. No, this is with the version in the HaikuPorts binary repo still; but based on the backtrace I'm guessing it's the BWebPage code to blame.

Patch-set: 1
Label: Code-Review=0