x-mail MIME: make Subject attribute editable
Fixes #2832

For details, please check the comments to ticket #2832.
In short: while not needed normally, it'd be nice to be able to edit
the subject attribute, for example when the sender accidentally forgot
to enter a subject or made an especially awkward typo that would spoil
querying for it.

Also: sentence casing the attribute name to "Reply to".