Update patch set 2
Patch Set 2:

> Nice, I think I was the one asking them to clarify memory fragmentation in their documentation.

Yes, I saw the Hacker News thread on it that you replied to :)

> Another interesting one is https://www.scott-a-s.com/projects/#streamflow and the one they compare against. But this one seems more ready to use.

Yes, that one looks unmaintained since 4 years ago, and not as well tested. The design does look a bit similar, but it seems to lack a global cache in favor of thread caches only; which is probably not a good idea for Haiku seeing how many cross thread allocations and deallocations we have due to the Be API threading model.

Patch-set: 2