Update patch set 2
Patch Set 2: Code-Review-2

It looks like the thread cleanup rpmalloc_thread_finalize() is never called on thread exit. This leads to leaking the thread local structures. This can be easily triggered by running an application and opening and closing menus for a couple of seconds (as that will create windows and hence spawn threads).

The same is true for the global finalize, although this is less useful since the whole team and its address space is going away anyway and the finalize would only add overhead.

I have locally introduced ways to call the thread finalize functions on thread exit to confirm that this is indeed a leak and not just memory not being returned to the OS. Getting this done properly is a bit tricky due to some exit work like cxa_finalize and pthread cleanup still needing a functional heap.

I've also played with porting the TLS usage from pthread keys to the simpler native TLS API and was able to measure a performance increase of about 5% on an uncached kernel build. Shall I post the patch here?

Porting from mmap to the more direct create_area is similarly simple and resulted in another 1% performance gain. The impact here is a lot smaller as these are a lot less frequent. A nice side effect of using create_area is that it allows to name the heap areas "heap area" again instead of the less obvious "mmap area" any mmap will get.

Patch-set: 2
Label: Code-Review=-2