rpmalloc: Use create_area instead of mmap.
It has lower overhead and allows to name the areas so that they are
labeled as heap areas.

Note that we can still use munmap for releasing. It has the nicer
interface of using the base address and length which we already know
instead of needing the area id that would have to be queried with an
additional syscall.