Update patch set 1
Patch Set 1:

Wouldn't a syslog be helpful when this happens? I would like to know when the system can't make sense of a partition and ignore it.

Also I should mention there are possibly weird things happening with multi session discs where ISO9660 is allowed to refer to things in a previous session, and we don't handle that properly. An easy test case is Audio CD with a data track at the end (including for example a video clip).

Patch-set: 1
Label: Code-Review=0