Update patch set 4
Patch Set 4:

> Would it make sense to have a rpmalloc branch? (Or do I need to learn more Gerrit magic?)

Gerrit keeps track of a sequence of commits without need for a branch (see the "related changes" box). If we start doing this a lot it would be a good practise to also tag the changes with a "topic", which is shown in Gerrit main page to more easily allow identifying commits there.

Patch-set: 4