Update patch set 1
Patch Set 1:

Hi Adrien,

> I don't think it would be possible to update the pen location properly. BPicture does not know the font metrics, it should just store commands to be replayed at a later time.
> 
> So, when recording we should not rely on anything from the current view state (pen location, font size, ...) and end up storing that data in the BPicture.
> 
> This probably means changing the app_server interface so the pen location is fully managed on app_server side, by introducing a new "draw at current pen location, whatever it is" call, and not reading the position from app_server and pushing it again.

at first my intuition was to follow your logic (if I understand correctly), but yesterday I've read through the documentation in the BeBook again. There is the general description in the Kits section, the class documentation and BView::DrawPicture() documentation. In all places it mentions that at the beginning of a BPicture recording, it will capture the complete drawing state.  That means a picture "knows" the font size and everything else even when these parameters are not set within the recording itself. A BPicture will always look exactly the same, all you can do is draw it at an offset, via the origin parameter of DrawPicture(). That in turn should mean that it is indeed correct to store absolute pen positions in the picture recording, as long as they are later relative to the origin for drawing the picture. What do you think? I have not yet investigated further as to what the actual bug is. I have a suspicion that when the BView marks the pen position as invalid, it may read from the wrong place inside app_server while a picture is being recorded. Why the MovePenBy() call would then "fix" that, does not fit that suspicion.

Patch-set: 1