Update patch set 1
Patch Set 1:

Indeed, that makes the problem simpler.

The core problem is that when we store a "draw text" in a BPicture, the pen position in the view state is not updated. But our code relies on reading it back from app_server, where the knowledge of font metrics reside and the pen position is normally set when drawing text.

However, the whole BPicture recording happens on app_server side (BView only sends AS_VIEW_BEGIN_PICTURE / AS_VIEW_END_PICTURE) so it would still be more correct to solve this problem on app_server side (update the pen position when drawing text, even while recording a picture)

We probably also need to check no other cases are missing, and we also correctly handle font changes (updating the font metrics when SetFont is called while recording a picture)

Patch-set: 1