Update patch set 1
Patch Set 1:

> Indeed, that makes the problem simpler.
> 
> The core problem is that when we store a "draw text" in a BPicture, the pen position in the view state is not updated. But our code relies on reading it back from app_server, where the knowledge of font metrics reside and the pen position is normally set when drawing text.
> 
> However, the whole BPicture recording happens on app_server side (BView only sends AS_VIEW_BEGIN_PICTURE / AS_VIEW_END_PICTURE) so it would still be more correct to solve this problem on app_server side (update the pen position when drawing text, even while recording a picture)
> 
> We probably also need to check no other cases are missing, and we also correctly handle font changes (updating the font metrics when SetFont is called while recording a picture)

After having studied the code, I am pretty sure that the problem is as you say: when performing a number of drawing operations, nothing actually happens other than storing the commands in the PictureDataWriter. For other commands, such as MovePenBy(), they are both recorded in the picture and applied on the server-side view. That is necessary, since PenLocation() should be usable while recording a picture and behave as expected. So the fix is indeed quite similar to your patch, only it needs to happen on the server side. I'll have to look at all drawing operations which affect pen location.

Independent from that fix, we should indeed have two separate DrawString() server messages, one for relative, one for absolute position. The additional PenLocation() which is used for the relative path on the client side, completely spoils the benefit of drawing asynchronously in the server, since that will have to wait for the server to actually have performed all the drawing so it can return the current pen location. I think the performance difference could affect text applications such as Gobe Productive.

Patch-set: 1